[tx-gen] Rename funding package and update examples for pyhmy version
