+++ test-automation/tx-gen/harmony_transaction_generator/__init__.py
@@ -20,7 +20,7 @@ from .account_manager import (
     return_balances,
     reset
 )
-from .setup import (
+from .fund import (
     fund_accounts
 )
 from .generator import (
